fix: change to component in page
