+++ pages/mydeck.tsx
@@ -1,14 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable @next/next/no-img-element */
 import type { NextPage } from "next";
-import { useDispatch, useSelector } from "react-redux";
-import { useEffect, useState } from "react";
-import {
-	detailPokemonURI,
-	getPokemonData,
-	setPokemonDetail,
-} from "../store/storeAction/pokemon";
-import { useRouter } from "next/router";
+import { useSelector } from "react-redux";
+import PokemonCard from "../components/PokemonCard";
 
 const MyDeck: NextPage = () => {
 	const pokemons = useSelector((state: IState) => state.pokemon.mydex);
@@ -21,35 +15,6 @@ const MyDeck: NextPage = () => {
 				{pokemons?.map((pokemon, index) => (
 					<PokemonCard key={index} {...pokemon} />
 				))}
-			</div>
-		</div>
-	);
-};
-
-const PokemonCard = (props: any) => {
-	const { name, url } = props;
-	const dispatch = useDispatch();
-
-	const route = useRouter();
-	const changeRoute = () => {
-		dispatch(setPokemonDetail(props));
-		route.push("detail");
-	};
-
-	return (
-		<div className="w-1/2 md:w-1/4 lg:w-1/6  flex justify-center">
-			<div
-				className="mb-4 flex flex-col items-center cursor-pointer"
-				onClick={changeRoute}
-			>
-				<img
-					alt="pokemon-picture"
-					className="object-contain"
-					src={props?.sprites?.front_default}
-				/>
-				<p className="capitalize font-bold bg-yellow-900 px-2 rounded text-white -mt-2 shadow-md">
-					{name}
-				</p>
 			</div>
 		</div>
 	);
